Add scalars for Date attributes annotated with TemporalType.TIMESTAMP TemporalType.TIME #111
