+++ graphql/src/main/java/io/jmix/graphql/GraphQLConfiguration.java
@@ -33,6 +33,7 @@ import io.jmix.graphql.limitation.LimitationProperties;
 import io.jmix.graphql.limitation.OperationRateLimitInstrumentation;
 import io.jmix.graphql.limitation.OperationRateLimitService;
 import io.jmix.graphql.schema.*;
+import io.jmix.graphql.schema.scalar.ScalarTypes;
 import io.jmix.graphql.security.SpecificPermissionInstrumentation;
 import io.jmix.graphql.spqr.SpqrSchemaGenerator;
 import io.leangen.geantyref.GenericTypeReflector;
@@ -111,6 +112,8 @@ public class GraphQLConfiguration {
     protected Messages messages;
     @Autowired
     protected JmixTypeInfoGenerator jmixTypeInfoGenerator;
+    @Autowired
+    protected ScalarTypes scalarTypes;
 
 
     @Bean
@@ -124,10 +127,8 @@ public class GraphQLConfiguration {
 
     @Bean
     public Jackson2ObjectMapperBuilderCustomizer jacksonCustomizer() {
-        return builder -> {
-            builder.serializerByType(Timestamp.class,
-                    new DateSerializer(false, new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss")));
-        };
+        return builder -> builder.serializerByType(Timestamp.class,
+                new DateSerializer(false, new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss")));
     }
 
     @Bean
@@ -142,7 +143,7 @@ public class GraphQLConfiguration {
         // filter
         types.addAll(filterTypesGenerator.generateFilterTypes());
         // scalars
-        types.addAll(Arrays.asList(Types.scalars));
+        types.addAll(scalarTypes.scalars());
         // permissions
         types.addAll(permissionTypesGenerator.generatePermissionTypes());
         // messages
