+++ graphql/src/main/java/io/jmix/graphql/schema/BaseTypesGenerator.java
@@ -16,30 +16,25 @@
 
 package io.jmix.graphql.schema;
 
-import graphql.Scalars;
 import graphql.schema.*;
 import io.jmix.core.MetadataTools;
 import io.jmix.core.metamodel.model.MetaProperty;
-import io.jmix.graphql.MetadataUtils;
 import io.jmix.graphql.NamingUtils;
-import io.jmix.graphql.datafetcher.GqlEntityValidationException;
-import io.jmix.graphql.schema.scalar.CustomScalars;
+import io.jmix.graphql.schema.scalar.ScalarTypes;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Nullable;
-import java.math.BigDecimal;
-import java.time.*;
-import java.util.Date;
-import java.util.UUID;
 
 @Component("gql_BaseTypesGenerator")
 public class BaseTypesGenerator {
 
     @Autowired
     private MetadataTools metadataTools;
+    @Autowired
+    ScalarTypes scalarTypes;
 
     /**
      * Shortcut for input value definition that has list type
@@ -116,7 +111,7 @@ public class BaseTypesGenerator {
 
     public String getFieldTypeName(MetaProperty metaProperty) {
         if (metaProperty.getType() == MetaProperty.Type.DATATYPE) {
-            return getScalarFieldTypeName(metaProperty);
+            return scalarTypes.getScalarFieldTypeName(metaProperty);
         }
 
         if (metaProperty.getType() == MetaProperty.Type.ENUM) {
@@ -131,74 +126,6 @@ public class BaseTypesGenerator {
         throw new UnsupportedOperationException(String.format("Can't define field type name for metaProperty %s class %s", metaProperty, metaProperty.getJavaType()));
     }
 
-
-    protected static String getScalarFieldTypeName(MetaProperty metaProperty) {
-        Class<?> javaType = metaProperty.getRange().asDatatype().getJavaClass();
-
-        // scalars from graphql-java
-
-        if (String.class.isAssignableFrom(javaType))
-            return Scalars.GraphQLString.getName();
-        if (Character.class.isAssignableFrom(javaType))
-            return Scalars.GraphQLChar.getName();
-        if (Integer.class.isAssignableFrom(javaType) || int.class.isAssignableFrom(javaType)) {
-            return Scalars.GraphQLInt.getName();
-        }
-        if (Short.class.isAssignableFrom(javaType) || short.class.isAssignableFrom(javaType)) {
-            return Scalars.GraphQLShort.getName();
-        }
-        if (Float.class.isAssignableFrom(javaType) || float.class.isAssignableFrom(javaType)
-                || Double.class.isAssignableFrom(javaType) || double.class.isAssignableFrom(javaType)) {
-            return Scalars.GraphQLFloat.getName();
-        }
-        if (Boolean.class.isAssignableFrom(javaType) || boolean.class.isAssignableFrom(javaType)) {
-            return Scalars.GraphQLBoolean.getName();
-        }
-
-        // more scalars added in jmix-graphql
-
-        if (UUID.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLUUID.getName();
-        }
-        if (Long.class.isAssignableFrom(javaType) || long.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLLong.getName();
-        }
-        if (BigDecimal.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLBigDecimal.getName();
-        }
-        if (Date.class.isAssignableFrom(javaType)) {
-            if (MetadataUtils.isDate(metaProperty)) {
-                return CustomScalars.GraphQLDate.getName();
-            }
-            if (MetadataUtils.isTime(metaProperty)) {
-                return CustomScalars.GraphQLTime.getName();
-            }
-            if (MetadataUtils.isDateTime(metaProperty)) {
-                return CustomScalars.GraphQLDateTime.getName();
-            }
-            throw new GqlEntityValidationException("Unsupported datatype mapping for date property " + metaProperty);
-        }
-        if (LocalDateTime.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLLocalDateTime.getName();
-        }
-        if (LocalDate.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLLocalDate.getName();
-        }
-        if (LocalTime.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLLocalTime.getName();
-        }
-        if (OffsetDateTime.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLOffsetDateTime.getName();
-        }
-        if (OffsetTime.class.isAssignableFrom(javaType)) {
-            return CustomScalars.GraphQLOffsetTime.getName();
-        }
-
-//        log.warn("getDatatypeFieldTypeName: can't resolve type for datatype meta property {} class {}", metaProperty, javaType);
-        // todo a couple of classes are not supported now
-        return "String";
-//        throw new UnsupportedOperationException(String.format("Can't define field type name for datatype class %s", javaType));
-    }
 
     protected String getReferenceTypeName(MetaProperty metaProperty) {
         if (metadataTools.isJpaEntity(metaProperty.getRange().asClass())) {
