+++ graphql/src/main/java/io/jmix/graphql/schema/scalar/CustomScalars.java
@@ -10,18 +10,14 @@ import java.text.SimpleDateFormat;
 
 public class CustomScalars {
 
-    public static final SimpleDateFormat SERIALIZATION_DATETIME_FORMAT = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS");
     public static final SimpleDateFormat SERIALIZATION_DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
     public static final SimpleDateFormat SERIALIZATION_TIME_FORMAT = new SimpleDateFormat("HH:mm:ss");
 
     public static GraphQLScalarType GraphQLDate = new DateScalar();
-    public static GraphQLScalarType GraphQLTime = new TimeScalar();
     public static GraphQLScalarType GraphQLDateTime = new DateTimeScalar();
     public static GraphQLScalarType GraphQLLocalDateTime = new LocalDateTimeScalar();
     public static GraphQLScalarType GraphQLLocalDate = new LocalDateScalar();
-    public static GraphQLScalarType GraphQLLocalTime = new LocalTimeScalar();
     public static GraphQLScalarType GraphQLOffsetDateTime = new OffsetDateTimeScalar();
-    public static GraphQLScalarType GraphQLOffsetTime = new OffsetTimeScalar();
 
     public static GraphQLScalarType GraphQLLong = GraphQLScalarType.newScalar()
             .name("Long")
