+++ graphql/src/main/java/io/jmix/graphql/schema/scalar/TimeScalar.java
@@ -1,46 +1,41 @@
 package io.jmix.graphql.schema.scalar;
 
-import graphql.schema.CoercingSerializeException;
+import graphql.schema.CoercingParseLiteralException;
 import graphql.schema.GraphQLScalarType;
-import io.jmix.graphql.schema.scalar.coercing.BaseDateCoercing;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import io.jmix.core.metamodel.datatype.Datatype;
+import io.jmix.graphql.schema.scalar.coercing.DatatypeCoercing;
 
 import java.sql.Time;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.util.Date;
+import java.text.ParseException;
+import java.time.LocalTime;
+import java.time.format.DateTimeParseException;
 
 
 public class TimeScalar extends GraphQLScalarType {
 
-    static final Logger log = LoggerFactory.getLogger(TimeScalar.class);
+    public TimeScalar(Datatype<LocalTime> localTimeDatatype, Datatype<Time> timeDatatype) {
+        super("Time", "Time type",
+                new DatatypeCoercing<Time>(Time.class, timeDatatype,
+                        "Please use the format 'HH:mm:ss' or 'HH:mm'") {
 
-    public TimeScalar() {
-        super("Time", "Time type", new BaseDateCoercing(LocalTimeScalar.LOCAL_TIME_FORMAT) {
-
-            @Override
-            public Object serialize(Object input) {
-                log.debug("serialize {}", input);
-
-                if (input instanceof Date) {
-                    Date date = (Date) input;
-                    return Instant.ofEpochMilli(date.getTime())
-                            .atZone(ZoneId.systemDefault())
-                            .toLocalTime().format(DateTimeFormatter.ISO_TIME);
-                }
-                throw new CoercingSerializeException(
-                        "Expected type 'Date' but was '" + input.getClass().getSimpleName() + "'.");
-
-            }
-
-            protected Object parseString(String value) {
-                // todo move formats to constant class
-                String timeStr = value.isEmpty() ? "00:00:00" : value;
-                return Time.valueOf(timeStr);
-            }
-        });
+                    // a bit more complicated logic, copied from Time parsing method of jmix-rest addon
+                    @Override
+                    public Time parseValue(Object input) {
+                        if (input instanceof String) {
+                            try {
+                                LocalTime localTime = localTimeDatatype.parse((String) input);
+                                if (localTime == null) {
+                                    return null;
+                                }
+                                return  Time.valueOf(localTime);
+                            } catch (DateTimeParseException | ParseException exception) {
+                                throw new CoercingParseLiteralException(incorrectFormatMessage);
+                            }
+                        }
+                        throw new CoercingParseLiteralException(
+                                "Expected type 'String' but was '" + input.getClass().getSimpleName() + "'.");
+                    }
+                });
     }
 
 }
