+++ graphql/src/test/groovy/io/jmix/graphql/schema/scalar/LocalDateScalarTest.groovy
@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-package io.jmix.graphql.schema.scalars
+package io.jmix.graphql.schema.scalar
 
 import graphql.language.StringValue
 import graphql.schema.Coercing
 import graphql.schema.CoercingParseLiteralException
 import graphql.schema.CoercingSerializeException
-import io.jmix.graphql.schema.scalar.OffsetDateTimeScalar
 import spock.lang.Specification
 
-import java.time.OffsetDateTime
+import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-class OffsetDateTimeScalarTest extends Specification {
+class LocalDateScalarTest extends Specification {
 
-    private final OffsetDateTimeScalar scalar = new OffsetDateTimeScalar()
+    private final LocalDateScalar scalar = new LocalDateScalar()
     private Coercing coercing
 
     @SuppressWarnings('unused')
@@ -36,12 +35,12 @@ class OffsetDateTimeScalarTest extends Specification {
         coercing = scalar.getCoercing()
     }
 
-    def "OffsetDateTime scalar test"() {
+    def "localDate scalar test"() {
         given:
-        def stringDate = new StringValue("2021-01-01T23:59:59+04:00")
-        def offsetDateTime = OffsetDateTime.from(
+        def stringDate = new StringValue("2021-01-01")
+        def localDate = LocalDate.from(
                 DateTimeFormatter
-                        .ISO_OFFSET_DATE_TIME
+                        .ISO_LOCAL_DATE
                         .parse(stringDate.getValue())
         )
         def parsedLiteral
@@ -51,30 +50,30 @@ class OffsetDateTimeScalarTest extends Specification {
         def nullParsedValue
 
         when:
-        parsedLiteral = (OffsetDateTime) coercing.parseLiteral(stringDate)
-        parsedValue = (OffsetDateTime) coercing.parseValue(stringDate.getValue())
-        serialized = coercing.serialize(offsetDateTime)
-        nullParsedLiteral = (OffsetDateTime) coercing.parseLiteral(new StringValue(""))
-        nullParsedValue = (OffsetDateTime) coercing.parseValue("")
+        parsedLiteral = (LocalDate) coercing.parseLiteral(stringDate)
+        parsedValue = (LocalDate) coercing.parseValue(stringDate.getValue())
+        serialized = coercing.serialize(localDate)
+        nullParsedLiteral = (LocalDate) coercing.parseLiteral(new StringValue(""))
+        nullParsedValue = (LocalDate) coercing.parseValue("")
 
         then:
-        parsedLiteral.isEqual(offsetDateTime)
-        parsedValue.isEqual(offsetDateTime)
+        parsedLiteral.isEqual(localDate)
+        parsedValue.isEqual(localDate)
         serialized == stringDate.getValue()
-        nullParsedLiteral.isEqual(OffsetDateTime.MIN)
-        nullParsedValue.isEqual(OffsetDateTime.MIN)
+        nullParsedLiteral.isEqual(LocalDate.MIN)
+        nullParsedValue.isEqual(LocalDate.MIN)
     }
 
-    def "OffsetDateTime scalar throws CoercingSerializeException"() {
+    def "localDate scalar throws CoercingSerializeException"() {
         when:
         coercing.serialize("")
 
         then:
         def exception = thrown(CoercingSerializeException)
-        exception.message == "Expected type 'OffsetDateTime' but was 'String'."
+        exception.message == "Expected type 'LocalDate' but was 'String'."
     }
 
-    def "OffsetDateTime scalar throws CoercingParseLiteralException with parseLiteral"() {
+    def "localDate scalar throws CoercingParseLiteralException with parseLiteral"() {
         when:
         coercing.parseLiteral("")
 
@@ -83,7 +82,7 @@ class OffsetDateTimeScalarTest extends Specification {
         exception.message == "Expected type 'StringValue' but was 'String'."
     }
 
-    def "OffsetDateTime scalar throws CoercingParseLiteralException with parseValue"() {
+    def "localDate scalar throws CoercingParseLiteralException with parseValue"() {
         when:
         coercing.parseValue(new StringValue(""))
 
@@ -92,12 +91,12 @@ class OffsetDateTimeScalarTest extends Specification {
         exception.message == "Expected type 'String' but was 'StringValue'."
     }
 
-    def "OffsetDateTime scalar throws CoercingParseLiteralException with wrong value"() {
+    def "localDate scalar throws CoercingParseLiteralException with wrong value"() {
         when:
         coercing.parseLiteral(new StringValue("1"))
 
         then:
         def exception = thrown(CoercingParseLiteralException)
-        exception.message == "Please use the format 'yyyy-MM-dd'T'HH:mm:ss+hh:mm'"
+        exception.message == "Please use the format 'yyyy-MM-dd'"
     }
 }
